Move intermediate nodes to schemas (importable without dependencies)
